Unified election information into one class.
